Represent thread lists as a bitfield
Storing thread lists as a vec of u32s was inefficient and we wasted a
lot of time on searching the lists every time we added a new thread.
I cobbled-together a bitfield implementation, which is a bit more
efficient.

Longer-term, I'll probably go back to a real data structure by the
standard library since, like a map, since the bitfield won't be helpful
for tracking matches.

But it was a fun diversion and gave me a chance to improve the
`ThreadList` struct and abstract away its implementation details.